feat: canonical-events-recording-004 - Converge Replay Projections and Enforce Recording Ownership
